requirements.txt
typer[all]
pdf2docx
python-docx
PyPDF2
Pillow
fpdf
